+++ CHANGELOG.md
@@ -16,6 +16,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   and the common-sense reasoning dataset Winogrande-sq.
 - Added the Dutch common sense reasoning dataset COPA-NL, which is part of the
 Dutch [DUMB benchmark](https://github.com/wietsedv/dumb)
+- Added new task for simplification and Dutch simplification dataset
+  [Duidelijke Taal](http://hdl.handle.net/10032/tm-a2-y8).
+  dataset.
 
 ## [v16.9.0] - 2025-12-16